requirements.txt
Django>=5.1.5
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Pillow>=10.0.0
djangorestframework>=3.15.2